fix commented out lines in c capture and fixes for clang

+++ src/image.cpp
@@ -202,7 +202,7 @@ namespace rsimpl
                 if(float depth = get_depth(depth_pixel_index))
                 {
                     // Determine the corresponding pixel location in our color image
-                    float depth_pixel[2] = {depth_x, depth_y}, depth_point[3], other_point[3], other_pixel[2];
+                    float depth_pixel[2] = {(float)depth_x, (float)depth_y}, depth_point[3], other_point[3], other_pixel[2];
                     rs_deproject_pixel_to_point(depth_point, &depth_intrin, depth_pixel, depth);
                     rs_transform_point_to_point(other_point, &depth_to_other, depth_point);
                     rs_project_point_to_pixel(other_pixel, &other_intrin, other_point);
